requirements.txt
boto3>=1.40.0,<2
jsonschema>=4.23.0,<5
openai>=1.30.0,<2
PyMuPDF>=1.24.0,<2
PyYAML>=6.0.0,<7
python-dotenv>=1.0.1,<2
